Set up CI with Azure Pipelines (#63)
* Set up CI with Azure Pipelines

* Exclude 1903,1909 from builds by default since agent can't support them

* Break out OS builds by agent image to help build OS-native images

+++ Developer Samples/WindowsContainers/Invoke-ContainerBuild.ps1
@@ -2,11 +2,13 @@
 param(
     [string][Parameter(Mandatory = $true)] $inruleReleaseTag,
     [string[]][Parameter(Mandatory = $false)] $imageTags = @(""),    
-    [string]$registry = "inrule"
+    [string]$registry = "inrule",
+    [string[]]$windowsServiceReleasesToUse = @("ltsc2016", "ltsc2019")
 )
 $ErrorActionPreference = "Stop"
 
 $baseImageTags = @()
+$serverTags = @()
 $catalogTags = @()
 $runtimeTags = @()
 $catManTags = @()
@@ -16,18 +18,19 @@ function Invoke-ContainerBuild {
     
     function buildServerImage {
         param($folder)
-        
+        if ($null -eq $folder) {
+            return $null
+        }
         try {
             Push-Location $folder
             $tag = "$registry/inrule-server:" + $folder.Name
             $cmd = "build --rm --no-cache -t $tag -f `"$folder\DOCKERFILE`" `"$folder`""
             runDockerBuild $cmd
-
+            $script:serverTags += $tag
             write-verbose "Built $tag"
         }
         finally {
-            Pop-Location      
-        
+            Pop-Location        
         }
         return $tag
     }
@@ -97,18 +100,18 @@ function Invoke-ContainerBuild {
         write-verbose $cmd
        
         if ($PSCmdlet.ShouldProcess($cmd) -and ($null -ne $cmd)) {
-          start-process "docker.exe" -ArgumentList $cmd  -NoNewWindow -Wait -Verbose
-          if ($LASTEXITCODE -ne 0) {
-              throw $Error[$Error.Count - 1]
-          }
+            start-process "docker.exe" -ArgumentList $cmd  -NoNewWindow -Wait -Verbose
         }
+    }
+    $windowsServiceReleasesToUse | foreach {
+        $folder = Get-item .\inrule-server\$_
+        buildServerImage $folder        
+        $script:baseImageTags += $_                
     }
 
     # we assume that all subfolders of the ..\WindowsContainers\inrule-server\**\* are build dirs
-    Get-ChildItem -Path .\inrule-server\ -Directory -Recurse | foreach {
-        $baseTag = buildServerImage $_        
-        $script:baseImageTags += $baseTag
-    }
+    
+       
 
     Write-Verbose "Base image tags: $baseImageTags"
     try {
@@ -125,7 +128,7 @@ function Invoke-ContainerBuild {
             Pop-Location
 
         }
-        $allTags = $baseImageTags + $catalogTags + $catManTags + $runtimeTags 
+        $allTags = $serverTags + $catalogTags + $catManTags + $runtimeTags 
         $allTags | Tee-Object .\tags-built.txt
     }
     finally {
